gateway/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.0
redis==5.2.0
pyyaml==6.0.2
pydantic==2.10.3
pyjwt==2.10.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0